Refactor: Add Checker interface, remove GameChecker
This update adds a Checker interface with an isGameOver method.
HorizontalChecker and VerticalChecker replace the GameChecker and each
only call isGameOver to check their win condition. Also replaced
GameChecker tests w/ HorizontalChecker tests and VerticalChecker tests.
