+++ src/main/java/com/connectfour/Checker.java
@@ -1,0 +1,5 @@
+package com.connectfour;
+
+public interface Checker {
+    boolean isGameOver(Board board);
+}
