+++ src/main/java/com/connectfour/VerticalChecker.java
@@ -1,0 +1,7 @@
+package com.connectfour;
+
+public class VerticalChecker implements Checker{
+    public boolean isGameOver(Board board) {
+        return false;
+    }
+}
